And develop

+++ android/core/model/src/main/java/com/sixkids/model/GroupSimple.kt
@@ -2,6 +2,6 @@ package com.sixkids.model
 
 data class GroupSimple(
     val headCount: Int,
-    val leaderId: Int,
+    val leaderId: Long,
     val students: List<Long>,
 )
